+++ src/torchmetrics/clustering/normalized_mutual_info_score.py
@@ -11,31 +11,33 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, List, Optional, Sequence, Union
+from typing import Any, List, Literal, Optional, Sequence, Union
 
 from torch import Tensor
 
-from torchmetrics.functional.clustering.mutual_info_score import mutual_info_score
-from torchmetrics.metric import Metric
+from torchmetrics.clustering.mutual_info_score import MutualInfoScore
+from torchmetrics.functional.clustering.normalized_mutual_info_score import (
+    _validate_average_method_arg,
+    normalized_mutual_info_score,
+)
 from torchmetrics.utilities.data import dim_zero_cat
 from torchmetrics.utilities.imports import _MATPLOTLIB_AVAILABLE
 from torchmetrics.utilities.plot import _AX_TYPE, _PLOT_OUT_TYPE
 
 if not _MATPLOTLIB_AVAILABLE:
-    __doctest_skip__ = ["MutualInfoScore.plot"]
+    __doctest_skip__ = ["NormalizedMutualInfoScore.plot"]
 
 
-class MutualInfoScore(Metric):
-    r"""Compute `Mutual Information Score`_.
+class NormalizedMutualInfoScore(MutualInfoScore):
+    r"""Compute `Normalized Mutual Information Score`_.
 
     .. math::
-        MI(U,V) = \sum_{i=1}^{|U|} \sum_{j=1}^{|V|} \frac{|U_i\cap V_j|}{N}
-        \log\frac{N|U_i\cap V_j|}{|U_i||V_j|}
+        NMI(U,V) = \frac{MI(U,V)}{M_p(U,V)}
 
-    Where :math:`U` is a tensor of target values, :math:`V` is a tensor of predictions,
-    :math:`|U_i|` is the number of samples in cluster :math:`U_i`, and :math:`|V_i|` is the number of samples in
-    cluster :math:`V_i`. The metric is symmetric, therefore swapping :math:`U` and :math:`V` yields the same mutual
-    information score.
+    Where :math:`U` is a tensor of target values, :math:`V` is a tensor of predictions, :math:`M_p(U,V)` is the
+    generalized mean of order :math:`p` of :math:`U` and :math:`V`, and :math:`MI(U,V)` is the mutual information score
+    between clusters :math:`U` and :math:`V`. The metric is symmetric, therefore swapping :math:`U` and :math:`V` yields
+    the same mutual information score.
 
     This clustering metric is an extrinsic measure, because it requires ground truth clustering labels, which may not
     be available in practice since clustering in generally is used for unsupervised learning.
@@ -47,44 +49,42 @@ class MutualInfoScore(Metric):
 
     As output of ``forward`` and ``compute`` the metric returns the following output:
 
-    - ``mi_score`` (:class:`~torch.Tensor`): A tensor with the Mutual Information Score
+    - ``nmi_score`` (:class:`~torch.Tensor`): A tensor with the Normalized Mutual Information Score
 
     Args:
+        average_method: Method used to calculate generalized mean for normalization
         kwargs: Additional keyword arguments, see :ref:`Metric kwargs` for more info.
 
     Example:
         >>> import torch
-        >>> from torchmetrics.clustering import MutualInfoScore
+        >>> from torchmetrics.clustering import NormalizedMutualInfoScore
         >>> preds = torch.tensor([2, 1, 0, 1, 0])
         >>> target = torch.tensor([0, 2, 1, 1, 0])
-        >>> mi_score = MutualInfoScore()
-        >>> mi_score(preds, target)
-        tensor(0.5004)
+        >>> nmi_score = NormalizedMutualInfoScore("arithmetic")
+        >>> nmi_score(preds, target)
+        tensor(0.4744)
 
     """
 
     is_differentiable: bool = True
-    higher_is_better: Optional[bool] = True
+    higher_is_better: Optional[bool] = None
     full_state_update: bool = False
     plot_lower_bound: float = 0.0
+    plot_upper_bound: float = 0.0
     preds: List[Tensor]
     target: List[Tensor]
     contingency: Tensor
 
-    def __init__(self, **kwargs: Any) -> None:
+    def __init__(
+        self, average_method: Literal["min", "geometric", "arithmetic", "max"] = "arithmetic", **kwargs: Any
+    ) -> None:
         super().__init__(**kwargs)
-
-        self.add_state("preds", default=[], dist_reduce_fx="cat")
-        self.add_state("target", default=[], dist_reduce_fx="cat")
-
-    def update(self, preds: Tensor, target: Tensor) -> None:
-        """Update state with predictions and targets."""
-        self.preds.append(preds)
-        self.target.append(target)
+        _validate_average_method_arg(average_method)
+        self.average_method = average_method
 
     def compute(self) -> Tensor:
-        """Compute mutual information over state."""
-        return mutual_info_score(dim_zero_cat(self.preds), dim_zero_cat(self.target))
+        """Compute normalized mutual information over state."""
+        return normalized_mutual_info_score(dim_zero_cat(self.preds), dim_zero_cat(self.target), self.average_method)
 
     def plot(self, val: Union[Tensor, Sequence[Tensor], None] = None, ax: Optional[_AX_TYPE] = None) -> _PLOT_OUT_TYPE:
         """Plot a single or multiple values from the metric.
@@ -106,8 +106,8 @@ class MutualInfoScore(Metric):
 
             >>> # Example plotting a single value
             >>> import torch
-            >>> from torchmetrics.clustering import MutualInfoScore
-            >>> metric = MutualInfoScore()
+            >>> from torchmetrics.clustering import NormalizedMutualInfoScore
+            >>> metric = NormalizedMutualInfoScore()
             >>> metric.update(torch.randint(0, 4, (10,)), torch.randint(0, 4, (10,)))
             >>> fig_, ax_ = metric.plot(metric.compute())
 
@@ -116,8 +116,8 @@ class MutualInfoScore(Metric):
 
             >>> # Example plotting multiple values
             >>> import torch
-            >>> from torchmetrics.clustering import MutualInfoScore
-            >>> metric = MutualInfoScore()
+            >>> from torchmetrics.clustering import NormalizedMutualInfoScore
+            >>> metric = NormalizedMutualInfoScore()
             >>> for _ in range(10):
             ...     metric.update(torch.randint(0, 4, (10,)), torch.randint(0, 4, (10,)))
             >>> fig_, ax_ = metric.plot(metric.compute())